+++ src/global/store/redux/reduxStore.tsx
@@ -8,7 +8,7 @@ import {
     storeSaga,
     storeSlice,
 } from 'src/pages/url/register/store/storeReducer'
-import { authSaga, authSlice } from 'src/features/auth/apis/authReducer'
+import { authSaga, authSlice } from 'src/features/auth/store/authReducer'
 
 const reducers = {
     routerReducer: routerSlice.reducer,
